portfolio-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
motor==3.3.2
pydantic==2.5.2
pydantic-settings==2.0.3
openai==1.3.5
anthropic==0.7.7
google-generativeai==0.3.1
python-multipart==0.0.6
python-magic==0.4.27
PyPDF2==3.0.1
